+++ grafana-plugin/src/utils/consts.ts
@@ -78,3 +78,10 @@ export const INCIDENT_HORIZONTAL_SCROLLING_STORAGE = 'isIncidentalTableHorizonta
 export const IRM_TAB = 'IRM';
 
 export const URL_REGEX = /^((https?|ftp|smtp):\/\/)?(www.)?[a-z0-9]+\.[a-z]+(\/[a-zA-Z0-9#]+\/?)*$/;
+
+export enum OnCallAGStatus {
+  Firing = 'Firing',
+  Resolved = 'Resolved',
+  Silenced = 'Silenced',
+  Acknowledged = 'Acknowledged',
+}
